Co-creation-projects/huailishang-AgentPlatformBase/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.0
python-dotenv>=1.0
requests>=2.31
